requirements.txt
requests
beautifulsoup4
pymongo
python-telegram-bot
apscheduler
fastapi
uvicorn